refactor: removed additional seed usage
