--- coins/core/src/crypto_address.rs
@@ -1,5 +1,0 @@
-/// The public address of a cryptocurrency.
-pub trait CryptoAddress {
-    /// Returns the public address as a string
-    fn public_address(&self) -> String;
-}
