+++ coins/core/src/lib.rs
@@ -7,13 +7,7 @@
 #![warn(missing_docs)]
 
 mod blockchain_connector;
-mod crypto_address;
-mod crypto_amount;
-mod crypto_wallet;
 mod error;
 pub mod prelude;
 pub use blockchain_connector::{BlockchainConnector, BlockchainConnectorBuilder, ConnectorType};
-pub use crypto_address::CryptoAddress;
-pub use crypto_amount::CryptoAmount;
-pub use crypto_wallet::{CryptoWallet, CryptoWalletBuilder};
 pub use error::Error;
