+++ coins/ethereum/src/lib.rs
@@ -124,9 +124,7 @@ mod error;
 pub use error::Error;
 pub use ethers;
 pub mod prelude;
-pub use walletd_coin_core::{
-    BlockchainConnector, CryptoAddress, CryptoAmount, CryptoWallet, CryptoWalletBuilder,
-};
+pub use walletd_coin_core::BlockchainConnector;
 pub use walletd_hd_key::{HDKey, HDNetworkType, HDPath, HDPathBuilder, HDPathIndex, HDPurpose};
 
 /// Represents the format of an Ethereum address (checksummed or non-checksummed)
