--- walletd/src/keystore.rs
@@ -1,2 +1,0 @@
-/// Stores the relevant keys and controls access to them
-pub struct KeyStore;
